+++ webapp/src/app/models/map-styles.ts
@@ -1,21 +1,13 @@
-export type MapStyles = {
-	walls: WallColors;
-} & {
-	[key: string]: MapStyle;
-};
+export interface MapStyles {
+	default: MapStyleType;
+	
+	[key: string]: MapStyleType | undefined;
+}
 
-export type MapStyle = {
-	sheet: string;
-	hasDoorMat?: boolean;
-} & {
-	[key: string]: any;
-};
+export interface MapStyleType {
+	[key: string]: MapStyle | undefined;
+}
 
-export interface WallColors {
-	blockFront: string;
-	blockTop: string;
-	pBlockFront: string;
-	pBlockTop: string;
-	npBlockFront: string;
-	npBlockTop: string;
+export interface MapStyle {
+	sheet?: string;
 }