+++ webapp/src/app/services/global-events.service.ts
@@ -22,9 +22,11 @@ export class GlobalEventsService {
 	offsetEntities = new Subject<Point>();
 	toggleVisibility = new Subject<void>();
 	showAddEntityMenu = new Subject<Point>();
+	updateEntities = new Subject<boolean>();
 	
 	updateCoords = new Subject<Point3 | undefined>();
 	updateTileSelectionSize = new Subject<Point | undefined>();
+	isAutotile = new BehaviorSubject(false);
 	showIngamePreview = new BehaviorSubject(false);
 	hasUnsavedChanges = new BehaviorSubject(false);
 	gridSettings = new BehaviorSubject<GridSettings>(Globals.gridSettings());