+++ webapp/src/app/services/phaser/entities/entity-manager.ts
@@ -83,6 +83,8 @@ export class EntityManager extends BaseObject {
 		this.visibilityKey = keyboard!.addKey(keyCodes.R, false);
 		
 		this.selectionBox = new SelectionBox(this.scene);
+		
+		Globals.globalEventsService.updateEntities.subscribe(() => this.resetEntities());
 	}
 	
 	
@@ -400,7 +402,10 @@ export class EntityManager extends BaseObject {
 			}
 			entities = (parsed as any[]).filter(v => this.isMapEntity(v));
 		} catch (e) {
-			Globals.snackbar.open('could not parse entities from clipboard', undefined, {duration: 2000});
+			Globals.snackbar.open('could not parse entities from clipboard', undefined, {
+				duration: 2000,
+				panelClass: 'snackbar-error'
+			});
 			return;
 		}
 		if (entities.length === 0) {
@@ -427,6 +432,19 @@ export class EntityManager extends BaseObject {
 			this.selectEntity(newEntity, entities.length > 1);
 		}
 		
+	}
+	
+	private async resetEntities() {
+		if (!this.map) {
+			return;
+		}
+		const exportedMap = this.map.exportMap();
+		await this.initialize(exportedMap, this.map);
+		if (this.active) {
+			for (const entity of this.entities) {
+				entity.setActive(true);
+			}
+		}
 	}
 	
 	deleteSelectedEntities() {