+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added overrideable JSON configs, see [#327](https://github.com/CCDirectLink/crosscode-map-editor/pull/327) for more details
+- Added most definitions to autotiles.json
+- Added new autotile type 4x4
+
 ## [1.7.1] 2024-07-21
 ### Fixed
 - Fixed Event editor not working